refactor: unify sign-up schema and implement step-based form with useFormContext
